[BE] refactor: 폴더구조 도메인형으로 변경
