+++ backend/driving-today/src/main/java/com/drivingtoday/domain/instructor/InstructorRepository.java
@@ -1,6 +1,6 @@
-package com.drivingtoday.repository;
+package com.drivingtoday.domain.instructor;
 
-import com.drivingtoday.entity.Instructor;
+import com.drivingtoday.domain.instructor.Instructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstructorRepository extends JpaRepository<Instructor, Long> {
