+++ backend/driving-today/src/main/java/com/drivingtoday/domain/review/ReviewRequest.java
@@ -1,4 +1,4 @@
-package com.drivingtoday.dto;
+package com.drivingtoday.domain.review;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
