+++ backend/driving-today/src/main/java/com/drivingtoday/domain/student/StudentRepository.java
@@ -1,6 +1,6 @@
-package com.drivingtoday.repository;
+package com.drivingtoday.domain.student;
 
-import com.drivingtoday.entity.Student;
+import com.drivingtoday.domain.student.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
